Extract bridge test fixtures into test_support module (#160)
